Artifacts from https://github.com/floodlight/loxigen/compare/1d8819a23e3decd195baa2d0306e9d887b17a50f...8dcf8a2618adfd827a20169c34401ed04a25d383
Loxigen Head commit floodlight/loxigen@8dcf8a2618adfd827a20169c34401ed04a25d383
commit 8dcf8a2618adfd827a20169c34401ed04a25d383
Merge: 1d8819a 99f1b43
Author: abat <abat@bigswitch.com>
Date:   Thu May 22 11:17:57 2014 -0700

    Merge into master from pull request #273:
    openflowj: fix u16obj read/write (https://github.com/floodlight/loxigen/pull/273)

commit 99f1b4343d77c44313e36880456a1d75e213eb1d
Author: Rich Lane <rlane@bigswitch.com>
Date:   Thu May 22 11:15:15 2014 -0700

    openflowj: fix u16obj read/write

    Fixes the bsn_tcp_flags OXM.

+++ openflowj/gen-src/main/java/org/projectfloodlight/openflow/protocol/ver12/OFOxmBsnTcpFlagsMaskedVer12.java
@@ -263,8 +263,8 @@ class OFOxmBsnTcpFlagsMaskedVer12 implements OFOxmBsnTcpFlagsMasked {
             int typeLen = bb.readInt();
             if(typeLen != 0x32104)
                 throw new OFParseError("Wrong typeLen: Expected=0x32104L(0x32104L), got="+typeLen);
-            U16 value = U16.of(bb.readByte());
-            U16 mask = U16.of(bb.readByte());
+            U16 value = U16.of(bb.readShort());
+            U16 mask = U16.of(bb.readShort());
 
             OFOxmBsnTcpFlagsMaskedVer12 oxmBsnTcpFlagsMaskedVer12 = new OFOxmBsnTcpFlagsMaskedVer12(
                     value,
@@ -303,8 +303,8 @@ class OFOxmBsnTcpFlagsMaskedVer12 implements OFOxmBsnTcpFlagsMasked {
         public void write(ChannelBuffer bb, OFOxmBsnTcpFlagsMaskedVer12 message) {
             // fixed value property typeLen = 0x32104L
             bb.writeInt(0x32104);
-            bb.writeByte(message.value.getRaw());
-            bb.writeByte(message.mask.getRaw());
+            bb.writeShort(message.value.getRaw());
+            bb.writeShort(message.mask.getRaw());
 
 
         }
